# generators/app/templates/src/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths
basePath: /
#
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json

paths:
  /person/{id}:
    x-swagger-router-controller: person
    get:
      description: Gets a specific person
      operationId: showPerson
      parameters:
        - name: id
          in: path
          description: id of user
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/PersonResponse"

        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

    put:
      description: upsert a person
      operationId: upsertPerson
      parameters:
        - name: id
          in: path
          description: id of user
          required: true
          type: string
        - in: body
          name: body
          description: Person object data to update
          required: true
          schema:
            $ref: '#/definitions/Person'
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/PersonResponse"

        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

    patch:
      description: partial update of a person
      operationId: patchPerson
      parameters:
        - name: id
          in: path
          description: id of user
          required: true
          type: string
        - name: JsonPatch
          in: body
          required: true
          schema:
            $ref: "#/definitions/PatchRequest"

      responses:
        "204":
          description: Success
          schema:
            $ref: "#/definitions/PersonResponse"

        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

    delete:
      description: delete a specific user
      operationId: destroyPerson
      parameters:
        - name: id
          in: path
          description: id of user
          required: true
          type: string
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
  /person:
    x-swagger-router-controller: person
    get:
      description: Returns a list of all persons
      operationId: indexPerson
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/PersonListResponse"

    post:
      description: Creates a person
      operationId: createPerson
      parameters:
        - in: body
          name: body
          description: Person object added to the databse
          required: true
          schema:
            $ref: '#/definitions/Person'

      responses:
        "201":
          description: Success
          schema:
            $ref: "#/definitions/PersonResponse"

        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  PersonListResponse:
    type: array
    items:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        description:
          type: string
        age:
          type: integer

  PersonResponse:
    required:
      - _id
    properties:
      _id:
        type: string
      name:
        type: string
      description:
        type: string
      age:
        type: integer

  PatchRequest:
    type: array
    items:
      $ref: "#/definitions/PatchDocument"

  PatchDocument:
    description: A JSONPatch document as defined by RFC 6902
    required:
     - "op"
     - "path"
    properties:
     op:
      type: string
      description: The operation to be performed
      enum:
       - "add"
       - "remove"
       - "replace"
       - "move"
       - "copy"
       - "test"
     path:
      type: string
      description: A JSON-Pointer
     value:
      type: object
      description: The value to be used within the operations.
     from:
      type: string
      description: A string containing a JSON Pointer value.

# Post definitions
  Person:
    title: A person
    type: object
    required:
      - name
      - description
      - age
    properties:
      name:
        type: string
      description:
        type: string
      age:
        type: integer

  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  SuccessResponse:
    required:
      - message
    properties:
      message:
        type: string